Image repository

See merge request eirsteir/team-14-software-engineering!24

+++ src/main/java/NTNU/IDATT1002/models/Image.java
@@ -27,13 +27,13 @@ public class Image {
 
 
   @NotBlank
-  private int imageID;
+  private Long imageID;
 
   @NotBlank
-  private int albumID;
+  private Long albumID;
 
   @NotBlank
-  private int metaDataID;
+  private Long metaDataID;
 
   @NotBlank
   @CreationTimestamp
@@ -45,7 +45,7 @@ public class Image {
   public Image() {
   }
 
-  public Image(int imageID, int albumID, int metaDataId, Date uploadAt, String path) {
+  public Image(Long imageID, Long albumID, Long metaDataId, Date uploadAt, String path) {
     this.imageID = imageID;
     this.albumID = albumID;
     this.metaDataID = metaDataId;
@@ -57,15 +57,15 @@ public class Image {
     this(image.getImageID(), image.getAlbumID(), image.getMetaDataID(), image.getUploadAt(), image.getPath());
   }
 
-  public void setImageID(int imageID) {
+  public void setImageID(Long imageID) {
     this.imageID = imageID;
   }
 
-  public void setAlbumID(int albumID) {
+  public void setAlbumID(Long albumID) {
     this.albumID = albumID;
   }
 
-  public void setMetaDataID(int metaDataID) {
+  public void setMetaDataID(Long metaDataID) {
     this.metaDataID = metaDataID;
   }
 
@@ -77,15 +77,15 @@ public class Image {
     this.path = path;
   }
 
-  public int getImageID() {
+  public Long getImageID() {
     return imageID;
   }
 
-  public int getAlbumID() {
+  public Long getAlbumID() {
     return albumID;
   }
 
-  public int getMetaDataID() {
+  public Long getMetaDataID() {
     return metaDataID;
   }
